Move BugRepro to top-level
